+++ src/server.cc
@@ -1689,8 +1689,12 @@ player_connected(Objid old_id, Objid new_id, bool is_newly_created)
         if (new_h->print_messages)
             send_message(new_h->listener, new_h->nhandle, "redirect_to_msg",
                          "*** Redirecting old connection to this port ***", 0);
-        network_close(existing_h->nhandle);
-        free_shandle(existing_h);
+        if (get_nhandle_refcount(existing_h->nhandle) > 1) {
+            existing_h->disconnect_me = true;
+        } else {
+            network_close(existing_h->nhandle);
+            free_shandle(existing_h);
+        }
         if (existing_listener == new_h->listener)
             call_notifier(new_id, new_h->listener, "user_reconnected");
         else {
@@ -1745,8 +1749,12 @@ player_switched(Objid old_id, Objid new_id, bool silent)
         if (!silent && new_h->print_messages)
             send_message(new_h->listener, new_h->nhandle, "redirect_to_msg",
                          "*** Redirecting old connection to this port ***", 0);
-        network_close(existing_h->nhandle);
-        free_shandle(existing_h);
+        if (get_nhandle_refcount(existing_h->nhandle) > 1) {
+            existing_h->disconnect_me = true;
+        } else {
+            network_close(existing_h->nhandle);
+            free_shandle(existing_h);
+        }
     } else {
         if (!silent && new_h->print_messages)
             send_message(new_h->listener, new_h->nhandle, "connect_msg",
@@ -2855,18 +2863,24 @@ name_lookup_callback(Var arglist, Var *ret, void *extra_data)
 {
     int nargs = arglist.v.list[0].v.num;
     Objid who = arglist.v.list[1].v.obj;
-    shandle *h = find_shandle(who);
     bool rewrite_connect_name = nargs > 1 && is_true(arglist.v.list[2]);
 
     network_handle nh;
     nh.ptr = extra_data;
 
-    if (!h || h->disconnect_me)
+    bool valid_connection;
+    {
+        std::lock_guard<std::recursive_mutex> lock(all_shandles_mutex);
+        shandle *h = find_shandle(who);
+        valid_connection = h && !h->disconnect_me.load();
+    }
+
+    if (!valid_connection)
         make_error_map(E_INVARG, "Invalid connection", ret);
     else
     {
         const char *name;
-        int status = lookup_network_connection_name(h->nhandle, &name);
+        int status = lookup_network_connection_name(nh, &name);
 
         /* If the server is shutting down, this is meaningless and creates
          * a bit of a mess anyway. So don't bother continuing. */